refact - fix oom issue
