+++ rust/src/lib.rs
@@ -78,7 +78,7 @@ pub unsafe extern "C" fn Java_com_ppaass_agent_vpn_LocalVpnService_onStartVpn(
 ) {
     let ppaass_vpn_server_config = PpaassVpnServerConfig::new();
     let android_logger_config = AndroidLoggerConfig::default()
-        .with_max_level(LevelFilter::Info)
+        .with_max_level(LevelFilter::Debug)
         .with_tag("PPAASS-VPN-RUST");
     android_logger::init_once(android_logger_config);
     // std::panic::set_hook(Box::new(|panic_info| {
